refactor(aihub): Consolidate memory management components and standardize UI for threads and agents
